[ul] Rearrange association module
- rename scu to client and scp to
   - more accurate names, since sometimes
     the nodes do not portray that role
   - rename ScuAssociation to ClientAssociation
   - rename ScpAssociation to ServerAssociation
- add root module documentation where missing
